Rewrite text domain check to actually work
The text domain checks we had before didn't really work at all. Bad
regex. Rewriting to use similar methods to the i18n core tools seems to
do the job, but makes it more difficult to point to specific lines of
code that contain issues. Still, it actually does work to detect
incorrect slugs and use of multiples, so it's better than before.